feature/login-as-index-page
Change login file to the index file
