+++ src/pages/index.tsx
@@ -1,123 +1,91 @@
-import Head from 'next/head'
-import Image from 'next/image'
-import { Inter } from 'next/font/google'
-import styles from '@/styles/Home.module.css'
-
-const inter = Inter({ subsets: ['latin'] })
+import React, { useState } from 'react';
+import { Form, Button, Card, Container, Row, Col } from 'react-bootstrap';
+import * as FaIcons from "react-icons/fa";
+import * as BsIcons from "react-icons/bs";
+import Image from 'next/image';
+import transLogo from 'src/images/transLogo.jpg';
+import bg from 'src/images/wbg.webp';
 
 export default function Home() {
+  // define state variables for email and password inputs
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+
+  // function to handle form submission
+  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    console.log('Email:', email);
+    console.log('Password:', password);
+  };
+
   return (
-    <>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main className={styles.main}>
-        <div className={styles.description}>
-          <p>
-            Get started by editing&nbsp;
-            <code className={styles.code}>src/pages/index.tsx</code>
-          </p>
-          <div>
-            <a
-              href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              By{' '}
-              <Image
-                src="/vercel.svg"
-                alt="Vercel Logo"
-                className={styles.vercelLogo}
-                width={100}
-                height={24}
-                priority
-              />
-            </a>
-          </div>
-        </div>
+    // Login form layout
 
-        <div className={styles.center}>
-          <Image
-            className={styles.logo}
-            src="/next.svg"
-            alt="Next.js Logo"
-            width={180}
-            height={37}
-            priority
-          />
-          <div className={styles.thirteen}>
-            <Image
-              src="/thirteen.svg"
-              alt="13"
-              width={40}
-              height={31}
-              priority
-            />
-          </div>
-        </div>
+    <div className="bg-light d-flex justify-content-center align-items-center vh-100 position-relative">
+      <div className={`position-absolute w-100 h-100`}>
+      <Image
+        src={bg}
+        alt='Wizeline Background'
+        layout="fill"
+        objectFit="cover"
+        quality={100}
+        loading="eager"
+        placeholder='blur'
+        style={{filter:'blur(10px)'}}
+      />
+      </div>
+      <div className="position-absolute w-100 h-100 d-flex justify-content-center align-items-center">
+        <Card style={{ width: '27rem' }}>
+          <Card.Body>
+          <div className="text-center pt-3 pb-4">
+            <Image 
+              src={transLogo} alt='WizeLine Logo' height={100}
+            /></div>
+            <Form onSubmit={handleSubmit}>
+              <Form.Group controlId="email">
+                <Form.Label>Email address</Form.Label>
+                <Form.Control
+                  type="email"
+                  placeholder="Enter email"
+                  value={email}
+                  onChange={(e) => setEmail(e.target.value)}
+                  className="form-control"
+                />
+              </Form.Group>
 
-        <div className={styles.grid}>
-          <a
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Docs <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Find in-depth information about Next.js features and&nbsp;API.
-            </p>
-          </a>
+              <Form.Group controlId="password">
+                <Form.Label>Password</Form.Label>
+                <Form.Control
+                  type="password"
+                  placeholder="Password"
+                  value={password}
+                  onChange={(e) => setPassword(e.target.value)}
+                  className="form-control"
+                />
+              </Form.Group>
 
-          <a
-            href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Learn <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Learn about Next.js in an interactive course with&nbsp;quizzes!
-            </p>
-          </a>
+              <Row className="mt-1 d-flex align-items-center">
+                <Col>
+                  <a href="#" className="a-login" style={{ fontSize: 'smaller' }}>Forgot your password?</a>
+                </Col>
+                <Col xs="auto">
+                  <Form.Check type="checkbox" style={{ fontSize: 'smaller' }} label="Remind me" className="pt-2" />
+                </Col>
+              </Row>
 
-          <a
-            href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Templates <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Discover and deploy boilerplate example Next.js&nbsp;projects.
-            </p>
-          </a>
+              <Row className="mt-3">
+                <Col>
+                  <Button variant="primary" type="submit" className="btn-primary w-100">
+                    <BsIcons.BsFillKeyFill className="mb-1" />
+                    &nbsp;&nbsp;Login
+                  </Button>
+                </Col>
+              </Row>
 
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Deploy <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Instantly deploy your Next.js site to a shareable URL
-              with&nbsp;Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-    </>
-  )
+            </Form>
+          </Card.Body>
+        </Card>
+      </div>
+    </div>
+  );
 }
